Changes: - Refactorized Meter class, removing unnecessary members, setting default values and checking values to set - Implemented but disabled/commented the DropDigits feature - Implemented but disabled/commented the ArtificialTimeSync feature - Increased app version to 1.39.05

+++ Xml/Port.cs
@@ -143,7 +143,7 @@ namespace Xml
         }
         
         [XmlIgnore]
-        public int MeterProtocol { get; set; }
+        public byte MeterProtocol { get; set; }
         
         [XmlIgnore]
         public int MeterLiveDigits { get; set; }
